refactor(server): remove the named property 'router http.ServeMux' and replaced it with http.Handler
PlayerServer now has all the methods that http.Handler has, which is
just ServeHTTP.

To fill in the http.Handler we assign it to router we create in
NewPlayerServer. We can do this because http.ServeMux has the method
ServeHTTP.

This let us remove our own ServeHTTP method, as we are already exposing
one via the embedded type.